+++ backend/canisters/user_index/impl/src/updates/wipe_legacy_unique_person_proofs.rs
@@ -1,11 +1,11 @@
 use crate::guards::caller_is_governance_principal;
 use crate::{RuntimeState, jobs, mutate_state};
-use canister_api_macros::update;
+use canister_api_macros::proposal;
 use canister_tracing_macros::trace;
 use tracing::info;
 use user_index_canister::wipe_legacy_unique_person_proofs::*;
 
-#[update(guard = "caller_is_governance_principal", msgpack = true)]
+#[proposal(guard = "caller_is_governance_principal")]
 #[trace]
 fn wipe_legacy_unique_person_proofs(_args: Args) -> Response {
     mutate_state(wipe_legacy_unique_person_proofs_impl)